test(ai): split peer-gap boundary tests under 1000 NCL (Refs #2509)
PR #2599 added three boundary tests for the EXPAND below-quota peer
3-province gap rule to `colonial_pressure_test.dart`. The combined file
ended at 1003 non-comment lines, tripping
`repo.dart_file_non_comment_line_size` (SPEC/program/dart-file-non-
comment-line-size.md) and failing the Quality (lint, analyze, observer)
job.

Moved the three boundary tests verbatim into a sibling test file
`colonial_pressure_peer_gap_boundary_test.dart` (new). The file imports
only what the boundary group needs (`colonizethis_test`,
`colonial_pressure`, `perception_snapshot`, `colonizethis_models`) and
preserves the SPEC reference comment that frames the rule. Left a
pointer comment in the original group so the split is discoverable.

No behavior change: same tests, same fixtures, same `expect` calls; only
file location moved.

Local verification:
- `dart run tool/check_dart_file_non_comment_line_size.dart` -- no
  violations (main file now 826 physical / well under 1000 NCL).
- `dart analyze test/colonial_pressure_test.dart
  test/colonial_pressure_peer_gap_boundary_test.dart` -- clean.
- `dart test test/colonial_pressure_test.dart
  test/colonial_pressure_peer_gap_boundary_test.dart` -- 25 / 25 pass
  (22 existing + 3 new boundary cases).
- `bash tool/check_test_imports.sh` -- pass.

Refs #2509
